Issue #841 Node Pools - try to reassign instances with matching pre-pulled images first, refactor AutoscaleManager
