--- api/src/main/java/com/epam/pipeline/manager/cluster/AutoscalerService.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2017-2019 EPAM Systems, Inc. (https://www.epam.com/)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.epam.pipeline.manager.cluster;
-
-import com.epam.pipeline.entity.configuration.PipelineConfiguration;
-import com.epam.pipeline.entity.pipeline.RunInstance;
-
-/**
- * Provides helper nethods for {@link AutoscaleManager}
- */
-public interface AutoscalerService {
-
-    boolean requirementsMatch(RunInstance instanceOld, RunInstance instanceNew);
-    RunInstance getDefaultInstance();
-    RunInstance configurationToInstance(PipelineConfiguration configuration);
-    RunInstance fillInstance(RunInstance instance);
-}
