+++ api/src/main/java/com/epam/pipeline/manager/cluster/autoscale/AutoscaleContants.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright 2017-2020 EPAM Systems, Inc. (https://www.epam.com/)
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.epam.pipeline.manager.cluster.autoscale;
+
+public final class AutoscaleContants {
+    public static final int NODEUP_SPOT_FAILED_EXIT_CODE = 5;
+    public static final int NODEUP_LIMIT_EXCEEDED_EXIT_CODE = 6;
+    public static final String NODE_POOL_PREFIX = "p-";
+
+    private AutoscaleContants() {
+        //no op
+    }
+}
