+++ api/src/main/resources/messages.properties
@@ -520,4 +520,4 @@ error.node.pool.missing.instance.type=Persistent node is missing 'instance type'
 error.node.pool.price.type.not.allowed=Price type ''{0}'' is not allowed.
 error.node.pool.instance.type.not.allowed=Instance type ''{0}'' is not allowed.
 error.node.pool.invalid.disk.size=Disk size shall be greater than zero.
-error.node.pool.invalid.count=Instance count shall be greater than zero.
+error.node.pool.invalid.count=Instance count shall be greater or equal to zero.
